backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
google-cloud-documentai>=2.20.0
google-cloud-aiplatform>=1.40.0
pytesseract>=0.3.10
pdf2image>=1.17.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
cachetools>=5.3.0
bleach>=6.1.0
pypdf>=4.1.0
python-magic>=0.4.27
slowapi>=0.1.9
google-cloud-secret-manager>=2.16.0
google-cloud-logging>=3.8.0
pytest-mock>=3.14.0
pytest-cov>=5.0.0
